[B] Assign header IDs to all converted docs
Fixes #1613
