+++ api/app/services/ingestions/converters/html.rb
@@ -1,7 +1,7 @@
 module Ingestions
   module Converters
     class Html < Ingestions::Converters::AbstractConverter
-
+      include Concerns::ConversionHelpers
       include Validator::StyleHelpers
 
       def perform
@@ -21,14 +21,6 @@ module Ingestions
         insert_style_tag!
         ensure_header_ids!
         document_parsed.to_s
-      end
-
-      def ensure_header_ids!
-        nodes = document_parsed.xpath("//h1 | //h2 | //h3 | //h4 | //h5 | //h6")
-        nodes.each_with_index do |node, index|
-          next if node["id"]
-          node["id"] = Digest::MD5.hexdigest(index.to_s + node.name + node.text)
-        end
       end
 
       def parser
