+++ api/app/services/ingestions/converters/latex.rb
@@ -3,6 +3,7 @@ require "pandoc-ruby"
 module Ingestions
   module Converters
     class Latex < Ingestions::Converters::AbstractConverter
+      include Concerns::ConversionHelpers
 
       def perform
         convert_to_html
@@ -13,7 +14,18 @@ module Ingestions
         %w(tex)
       end
 
+      private
+
       def convert_to_html
+        ensure_header_ids!
+        document_parsed.to_s
+      end
+
+      def document_parsed
+        @document_parsed ||= Nokogiri::HTML raw_html
+      end
+
+      def raw_html
         PandocRuby.latex(contents, :s).to_html
       end
